More changes to do with uint64_t. Important changes in pstDtToRung, and in pstDomainOrder. Cosmetic changes to printfs to use %llu where uint64_t is passed. Removed the dEmax, dWmax, etc counters from the gravity code now. These are redundant since we now have per processor output of timings and other information.

+++ master.h
@@ -60,7 +60,7 @@ typedef struct msrContext {
     /*
      * Domain Decomposition Done
      */
-    int *nRung;
+    uint64_t *nRung;
     int bDoneDomainDecomp;
     int iLastRungDD;
     int iLastRungSD;
@@ -195,7 +195,7 @@ void msrSetRung(MSR msr, int iRung);
 int msrMaxRung(MSR msr);
 
 void msrSwitchTheta(MSR msr,double);
-int msrMaxOrder(MSR msr);
+uint64_t msrMaxOrder(MSR msr);
 
 void msrInitTimeSteps(MSR,double,double);
